blogtng_comment move to new class/file Comment
