+++ tpl/default/comments.php
@@ -6,8 +6,9 @@
  * comments. It is used to display a single comment in
  * the list and is called multiple times (once for each shown comment)
  *
- * @var blogtng_comment $comment
+ * @var dokuwiki\plugin\blogtng\entities\Comment $comment
  */
+
 ?>
 <div class="blogtng_comment blogtng_comment_status_<?php $comment->tpl_status()?>" id="comment_<?php $comment->tpl_cid()?>">
     <img src="<?php $comment->tpl_avatar(48,48)?>" class="avatar" width="48" height="48" alt="" align="left" />
